MAINT: Factor out PRLStructure into its own module
