+++ prlworkflows/__init__.py
@@ -3,4 +3,4 @@ from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
 
-from prlworkflows.structure_builders.sqs import SQS, enumerate_sqs
+from prlworkflows.structure_builders import PRLStructure
