+++ prlworkflows/structure_builders/__init__.py
@@ -1,3 +1,5 @@
 """Functions to build Structures"""
 
+from .prl_structure import PRLStructure
+from .sqs import enumerate_sqs
 from .sqs_db import SQSDatabase, get_structures_from_database
