+++ prlworkflows/structure_builders/dilute.py
@@ -1,0 +1,4 @@
+"""Convienence functions for building dilute structures"""
+
+def dilute_substitution(structure, sublattice, equivalent_sublattices=None, equivalent_wyckoff_sites=None):
+    pass
