+++ prlworkflows/structure_builders/sqs.py
@@ -7,11 +7,14 @@ These species in pymatgen Structures are named to `Xab`, which corresponds to at
 
 from __future__ import division
 
-import itertools
 import copy
-import numpy as np
-from pymatgen import Structure
+import itertools
+
 import pymatgen as pmg
+from pymatgen import Structure
+
+from .prl_structure import PRLStructure
+
 
 class AbstractSQS(Structure):
     """A pymatgen Structure with special features for SQS.
@@ -111,7 +114,7 @@ class AbstractSQS(Structure):
         site_ratios = [sum(ratios) for ratios in self.sublattice_site_ratios]
 
         # create the SQS and add all of these properties to our SQS
-        concrete_sqs = SQS.from_sites(self_copy.sites)
+        concrete_sqs = PRLStructure.from_sites(self_copy.sites)
         concrete_sqs.sublattice_configuration = sublattice_configuration
         concrete_sqs.sublattice_occupancies = sublattice_occupancies
         concrete_sqs.sublattice_site_ratios = site_ratios
@@ -161,130 +164,6 @@ class AbstractSQS(Structure):
         return sqs
 
 
-class SQS(Structure):
-    """A pymatgen Structure with special features for SQS.
-    """
-
-    def __init__(self, *args, **kwargs):
-        """Create a SQS object
-
-        Parameters
-        ----------
-        args :
-            args to pass to Structure
-        sublattice_configuration : [[str]]
-            Sublattice configuration  e.g. `[['Fe', 'Ni'], ['Fe']]`.
-        sublattice_occupancies : [[float]]
-            Fraction of the sublattice each element in the configuration  has e.g. `[[0.3333, 0.6666], [1]]`.
-        sublattice_site_ratios : [float]
-            Ratios of sublattice multiplicity  e.g. `[3, 1]`.
-        kwargs :
-            kwargs to pass to Structure
-        """
-        self.sublattice_configuration = kwargs.pop('sublattice_configuration', None)
-        self.sublattice_occupancies = kwargs.pop('sublattice_occupancies', None)
-        self.sublattice_site_ratios = kwargs.pop('sublattice_site_ratios', None)
-        super(SQS, self).__init__(*args, **kwargs)
-
-    def __eq__(self, other):
-        """
-        self and other are equivalent if the sublattice models are equal
-
-        Parameters
-        ----------
-        other : SQS
-        """
-        if not isinstance(other, SQS):
-            return False
-        subl_config = self.sublattice_configuration == other.sublattice_configuration
-        subl_site_ratios = self.sublattice_site_ratios == other.sublattice_site_ratios
-        subl_occupancies = self.sublattice_occupancies == other.sublattice_occupancies
-        return subl_config and subl_site_ratios and subl_occupancies
-
-    @property
-    def espei_sublattice_configuration(self):
-        """
-        Return ESPEI-formatted sublattice model [['a', 'b'], 'a'] for the concrete case
-        """
-        # short function to convert [['A', 'B'], ['A']] to [['A', 'B'], 'A'] as in ESPEI format
-        canonicalize_sublattice = lambda sl: sl[0] if len(sl) == 1 else sl
-        return [canonicalize_sublattice(sl) for sl in self.sublattice_configuration]
-
-    @property
-    def espei_sublattice_occupancies(self):
-        """
-        Return ESPEI-formatted sublattice occupancies [[0.3333, 0.6666], 1] for the concrete case
-        """
-        # short function to convert [[0.3333, 0.6666], [1]] to [[0.3333, 0.6666], 1] as in ESPEI format
-        canonicalize_sublattice = lambda sl: sl[0] if len(sl) == 1 else sl
-        return [canonicalize_sublattice(sl) for sl in self.sublattice_occupancies]
-
-    def as_dict(self, verbosity=1, fmt=None, **kwargs):
-        d = super(SQS, self).as_dict(verbosity=verbosity, fmt=fmt, **kwargs)
-        d['sublattice_configuration'] = self.sublattice_configuration
-        d['sublattice_occupancies'] = self.sublattice_occupancies
-        d['sublattice_site_ratios'] = self.sublattice_site_ratios
-        return d
-
-    @classmethod
-    def from_dict(cls, d, fmt=None):
-        sqs = super(SQS, cls).from_dict(d, fmt=fmt)
-        sqs.sublattice_configuration = d.get('sublattice_configuration')
-        sqs.sublattice_occupancies = d.get('sublattice_occupancies')
-        sqs.sublattice_site_ratios = d.get('sublattice_site_ratios')
-        return sqs
-
-    @staticmethod
-    def reindex_sublattice(new_indices, subl_model, subl_occupancies, subl_site_ratios):
-        """
-        Re-index the passed sublattice model, occupancies and site ratios according to the new index.
-
-        Parameters
-        ----------
-        new_indices : [int]
-            List of indicies corresponding to sublattices. There should be no duplicates. Specifically,
-            sorted(new_indices) == list(range(len(subl_model))
-        subl_model : [[str]]
-            Sublattice configuration  e.g. `[['Fe', 'Ni'], ['Fe']]`.
-        subl_occupancies : [[float]]
-            Fraction of the sublattice each element in the configuration  has e.g. `[[0.3333, 0.6666], [1]]`.
-        subl_site_ratios : [float]
-            Ratios of sublattice multiplicity  e.g. `[3, 1]`.
-
-        Returns
-        -------
-        tuple
-            Tuple of (sublattice model, occupancies, site ratios) that have been re-indexed
-
-        Examples
-        --------
-
-        >>> SQS.reindex_sublattice([1, 0], [['Al', 'Ni'], ['Al']], [[0.333, 0.666], [1]], [3, 1])
-        ([['Al'], ['Al', 'Ni']], [[1], [0.333, 0.666]], [1, 3])
-        """
-        if sorted(new_indices) != list(range(len(subl_model))):
-            raise ValueError('Passed re-indexing indicies ({}) do not match the sublattice model indices ({}).'.format(new_indices, list(range(len(subl_model)))))
-        new_subl_model = [subl_model[i] for i in new_indices]
-        new_subl_occupancies = [subl_occupancies[i] for i in new_indices]
-        new_subl_site_ratios = [subl_site_ratios[i] for i in new_indices]
-        return (new_subl_model, new_subl_occupancies, new_subl_site_ratios)
-
-
-    def reindex(self, new_indices):
-        """
-        Re-index the instance sublattice model, occupancies and site ratios according to the new index.
-
-        Parameters
-        ----------
-        new_indices : [int]
-            List of indicies corresponding to sublattices. There should be no duplicates. Specifically,
-            sorted(new_indices) == list(range(len(subl_model))
-
-        """
-        self.sublattice_configuration, self.sublattice_occupancies, self.sublattice_site_ratios = \
-            SQS.reindex_sublattice(new_indices, self.sublattice_configuration, self.sublattice_occupancies, self.sublattice_site_ratios)
-
-
 def enumerate_sqs(structure, subl_model, scale_volume=True, skip_on_failure=False):
     """
     Return a list of all of the concrete Structure objects from an abstract Structure and concrete sublattice model.
@@ -305,8 +184,8 @@ def enumerate_sqs(structure, subl_model, scale_volume=True, skip_on_failure=Fals
 
     Returns
     -------
-    [SQS]
-        List of all concrete SQS objects that can be created from the sublattice model.
+    [PRLStructure]
+        List of all concrete PRLStructure objects that can be created from the sublattice model.
     """
     if len(subl_model) != len(structure.sublattice_model):
         raise ValueError('Passed sublattice model ({}) does not agree with the passed structure ({})'.format(subl_model, structure.sublattice_model))
